refactor : mvc 피드백 받아 수정
