+++ src/main/java/roomescape/reservation/controller/ReservationApiController.java
@@ -1,0 +1,48 @@
+package roomescape.reservation.controller;
+
+import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.reservation.Reservation;
+import roomescape.reservation.dto.AddReservationRequest;
+import roomescape.reservation.exception.MissingRequiredFieldException;
+import roomescape.reservation.ReservationService;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+@RequestMapping("/reservations")
+public class ReservationApiController {
+    private final ReservationService reservationService;
+
+    public ReservationApiController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping
+    @ResponseBody
+    public ResponseEntity<List<Reservation>> getReservations() {
+        List<Reservation> reservations = reservationService.getAllReservations();
+        return ResponseEntity.ok().body(reservations);
+    }
+
+    @PostMapping
+    public ResponseEntity<?> addReservation(@RequestBody @Valid AddReservationRequest reservationRequest) {
+        String name = reservationRequest.getName();
+        String date = reservationRequest.getDate();
+        String time = reservationRequest.getTime();
+
+        Reservation reservation = reservationService.addReservation(name, date, time);
+        URI location = URI.create("/reservations/" + reservation.getId());
+
+        return ResponseEntity.created(location).body(reservation);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> deleteReservation(@PathVariable Long id) {
+        reservationService.cancelReservation(id);
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).body("예약이 취소되었습니다.");
+    }
+}
