+++ src/main/java/roomescape/reservation/dto/AddReservationRequest.java
@@ -1,9 +1,14 @@
 package roomescape.reservation.dto;
 
+import jakarta.validation.constraints.NotBlank;
+
 public class AddReservationRequest {
-    private String name;
-    private String date;
-    private String time;
+    @NotBlank(message = "이름은 필수 값입니다.")
+    final private String name;
+    @NotBlank(message = "날짜은 필수 값입니다.")
+    final private String date;
+    @NotBlank(message = "시간은 필수 값입니다.")
+    final private String time;
 
     public AddReservationRequest(String name, String date, String time) {
         this.name = name;
